backend/requirements.txt
flask
flask-cors
openai
beautifulsoup4
requests
firebase-admin
python-dotenv
dnspython
pytest